About our copyright claim
All of the material in this site is copyrighted.  We are often asked how we can claim a copyright in a public domain book like The Bible.  While it is quite true that no one can have a copyright that excludes all others from preparing their own unique copy of the Bible or other public domain works for copyright protection, our Bibles and other materials are not exactly like any others and are fully protected by copyright laws in all countries.  If you want to prepare your own unique Bible or other Torah materials for the Web, we are open to giving help and advice toward that end.

You may use materials from this site only for your private study and teaching, not for publishing in any form, including any Web site.  Under our liberal understanding of the Fair Use Doctrine, you may use a few short quotations from our site in any other work, including copyrighted publications.  Any other use requires written permission from us.

We expressly invite links to any of the text files on our site (but not to zips or graphics, for which you should link to our pages that link to them).  We provide detailed information on how to link to individual paragraphs and verses elsewhere on this site.

This site has been here since 24 Av 5758 (16 August 1998).